requirements.txt
argparse==1.4.0
rapidfuzz
gensim==3.5.0
nltk==3.3
numpy==1.14.5
PyYAML==5.1
SPARQLWrapper==1.8.2
torch==0.4.1
